Move autotranslate models to rc-models
