+++ packages/rocketchat-autotranslate/server/index.js
@@ -1,5 +1,3 @@
-import './models/Messages';
-import './models/Subscriptions';
 import './settings';
 import './permissions';
 import './autotranslate';
