--- packages/rocketchat-autotranslate/server/models/Messages.js
@@ -1,19 +1,0 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
-
-RocketChat.models.Messages.addTranslations = function(messageId, translations) {
-	const updateObj = {};
-	Object.keys(translations).forEach((key) => {
-		const translation = translations[key];
-		updateObj[`translations.${ key }`] = translation;
-	});
-	return this.update({ _id: messageId }, { $set: updateObj });
-};
-
-RocketChat.models.Messages.addAttachmentTranslations = function(messageId, attachmentIndex, translations) {
-	const updateObj = {};
-	Object.keys(translations).forEach((key) => {
-		const translation = translations[key];
-		updateObj[`attachments.${ attachmentIndex }.translations.${ key }`] = translation;
-	});
-	return this.update({ _id: messageId }, { $set: updateObj });
-};
